Deduplicate repl history
The easiest way is to not add the line if it's the same as the last
one, a more advanced one would be to locate other identical lines at
an earlier position, remove them from there and adding them at the
top.  Finally, deduplicating the history file upon loading it would be
the best.